Devel v3.0

+++ build/build-brics.sh
@@ -1,14 +1,21 @@
 #!/bin/bash
 
 # Building SetupBRICS
+
 if [ ! -f lib/cdk-1.4.19.jar ]; then
 	echo "Failed to create SetupBRICS.jar. Cannot locate cdk-1.4.19.jar in ../lib"
     exit -1
 fi
 
+if [ ! -f DENOPTIM-GUI.jar ]; then
+    echo "Failed to locate DENOPTIM-GUI.jar"
+    exit -1
+fi
+cp DENOPTIM-GUI.jar lib/
+
 
 find ../src/misc/RDKITFragmenter/src/ -name *.java > javafiles.txt
-javac -cp lib/cdk-1.4.19.jar:lib/DENOPTIM.jar @javafiles.txt -encoding utf-8 -d .
+javac -cp lib/cdk-1.4.19.jar:lib/DENOPTIM-GUI.jar @javafiles.txt -encoding utf-8 -d .
 
 if [ "$?" != "0" ]; then
     rm javafiles.txt
@@ -21,7 +28,7 @@ rm javafiles.txt
 
 echo "Manifest-Version: 1.0" > manifest.mf
 echo "Main-Class: setupbrics.SetupBRICS" >> manifest.mf
-echo "Class-Path: lib/cdk-1.4.19.jar lib/DENOPTIM.jar" >> manifest.mf
+echo "Class-Path: lib/cdk-1.4.19.jar lib/DENOPTIM-GUI.jar" >> manifest.mf
 echo >> manifest.mf
 
 jar cvfm SetupBRICS.jar manifest.mf setupbrics
